move perf event
